+++ src/_examples/characterize_openai.py
@@ -2,6 +2,7 @@ import os
 
 from pydantic import BaseModel, Field
 
+from mojentic.llm.gateways.models import LLMMessage, MessageRole
 from mojentic.llm.gateways.openai import OpenAIGateway
 
 api_key = os.getenv("OPENAI_API_KEY")
@@ -14,8 +15,8 @@ class Feeling(BaseModel):
 
 response = gateway.complete(
     model="gpt-4o-mini",
-    messages=[{'role': 'user', 'content': "Hello, how are you?"}],
-    response_model=Feeling,
+    messages=[LLMMessage(content="Hello, how are you?")],
+    object_model=Feeling,
     temperature=1.0,
     num_ctx=32768,
     num_predict=-1