+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added support for image file paths in OllamaGateway's complete and complete_stream methods
+
 ## [0.2.6] - 2025-03-09
 
 ### Added